기본 기능 구현 완료

+++ controllers/videoControllers.js
@@ -1,6 +1,7 @@
 // import {videos} from "../db"
 import routes from '../routes';
 import Video from '../models/Video';
+import Comment from '../models/Comment';
 
 export const home = async (req, res) => {
   try {
@@ -53,8 +54,9 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id).populate('creator');
-    console.log(video);
+    const video = await Video.findById(id)
+    // eslint-disable-next-line indent
+    .populate('creator').populate('comments');
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -68,8 +70,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    console.log(video);
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -96,10 +101,55 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
+  const video = await Video.findById(id);
   try {
-    await Video.findOneAndRemove({ _id: id });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
+};
+
+export const postRegisterView = async (req, res) => {
+  const {
+    params: { id },
+  } = req;
+
+  try {
+    const video = await Video.findById(id);
+    video.views += 1;
+    video.save();
+    res.status(200);
+  } catch (error) {
+    res.status(400);
+    res.end();
+  } finally {
+    res.end();
+  }
+};
+
+export const postAddComment = async (req, res) => {
+  const {
+    params: { id },
+    body: { comment },
+    user,
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    const newComment = await Comment.create({
+      text: comment,
+      create: user.id,
+    });
+    // console.log('newComment', newComment);
+    video.comments.push(newComment.id);
+    video.save();
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
 };
